requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
psycopg[binary,pool]==3.2.1
python-multipart==0.0.9
redis==5.0.7
celery==5.4.0
Pillow==10.4.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
asgi-lifespan==2.1.0
openai==1.52.2
pgvector==0.2.4
boto3==1.35.35
PyJWT==2.9.0
argon2-cffi==23.1.0
google-auth==2.35.0
python-jose==3.3.0
requests==2.32.3
email-validator==2.2.0
torch==2.4.0
open-clip-torch==2.24.0
opencv-python-headless
scikit-image
numpy
